pyproject.toml
[project]
name = "gavellm"
version = "0.1.0"
description = "Open source LLM-as-a-judge"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "openai>=1.106.1",
    "python-dotenv>=1.1.1",
    "jinja2>=3.1.4",
    "pyyaml>=6.0.2",
    "jsonschema>=4.21.1",
    "typer>=0.12.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
